refactor: Organize components into subdirectories by page
Refactored the 'components' folder to improve code organization. Components are now grouped into subdirectories based on the pages they are used on.

This change enhances maintainability and improves code structure.
