+++ renderer/components/home/AddModal.jsx
@@ -1,7 +1,7 @@
 import electron from "electron";
 import { useState } from "react";
 import { Book } from "./Book";
-import { useStore } from "../store/store";
+import { useStore } from "../../store/store";
 import { useNotification } from "doom-react-notifications";
 
 const ipcRenderer = electron.ipcRenderer || false;
@@ -171,9 +171,9 @@ export const AddModal = ({ trigger_id }) => {
               // TODO: find a better way to do this, and add a information message.
               disabled={
                 book.title != "" &&
-                  book.author != "" &&
-                  book.path != "" &&
-                  book.category != ""
+                book.author != "" &&
+                book.path != "" &&
+                book.category != ""
                   ? false
                   : true
               }
