+++ renderer/components/home/EditModal.jsx
@@ -1,6 +1,6 @@
 import electron from "electron";
 import { shallow } from "zustand/shallow";
-import { useStore } from "../store/store";
+import { useStore } from "../../store/store";
 import { Book } from "./Book";
 import { useNotification } from "doom-react-notifications";
