+++ renderer/pages/config.jsx
@@ -4,6 +4,7 @@ import {
   AiOutlineBook,
   AiOutlineUser,
 } from "react-icons/ai";
+import { Theme } from "../components/config/Theme";
 
 function Config() {
   return (
@@ -34,20 +35,37 @@ function Config() {
               <h3 className="text-2xl font-bold uppercase text-primary">
                 Dashboard config
               </h3>
-              <p>
-                Lorem ipsum dolor sit amet, officia excepteur ex fugiat
-                reprehenderit enim labore culpa sint ad nisi Lorem pariatur
-                mollit ex esse exercitation amet. Nisi anim cupidatat excepteur
-                officia. Reprehenderit nostrud nostrud ipsum Lorem est aliquip
-                amet voluptate voluptate dolor minim nulla est proident. Nostrud
-                officia pariatur ut officia. Sit irure elit esse ea nulla sunt
-                ex occaecat reprehenderit commodo officia dolor Lorem duis
-                laboris cupidatat officia voluptate. Culpa proident adipisicing
-                id nulla nisi laboris ex in Lorem sunt duis officia eiusmod.
-                Aliqua reprehenderit commodo ex non excepteur duis sunt velit
-                enim. Voluptate laboris sint cupidatat ullamco ut ea consectetur
-                et est culpa et culpa duis.
-              </p>
+              <div className="w-full max-w-xs form-control">
+                <label className="label">
+                  <span className="label-text">theme:</span>
+                </label>
+                <Theme />
+              </div>
+              <div className="w-full max-w-xs form-control">
+                <label className="label">
+                  <span className="label-text">pdf Reader:</span>
+                </label>
+                <input
+                  id="pdf-reader"
+                  type="text"
+                  name="pdf-reader"
+                  className="w-full max-w-xs input input-bordered"
+                  placeholder="zathura"
+                />
+              </div>
+              <div className="w-full max-w-xs form-control">
+                <label className="label">
+                  <span className="label-text">sort books by:</span>
+                </label>
+                <select className="select select-bordered" defaultValue="title">
+                  <option value="title">title</option>
+                  <option value="year">year</option>
+                  <option value="author">author</option>
+                </select>
+              </div>
+              <button className="mt-5 btn btn-wide" disabled>
+                Save
+              </button>
             </div>
           </div>
         </div>
